feat(MyApp): final version, state hoisting to function parameter
